Add called_n_times(n)
Matcher that will verify a call has been called the expected number of
times.

'Now did the Lord say, "First thou pullest the Holy Pin. Then thou must
count to three. Three shall be the number of the counting and the number
of the counting shall be three. Four shalt thou not count, neither shalt
thou count two, excepting that thou then proceedeth to three. Five is
right out. Once the number three, being the number of the counting, be
reached, then lobbest thou the Holy Hand Grenade in the direction of
thine foe, who, being naughty in my sight, shall snuff it."'